+++ Renderer/Mesh.hpp
@@ -1,6 +1,8 @@
 #pragma once
 #include <vector>
 #include "GLM/vec3.hpp"
+#include "GLFW/glfw3.h"
+#include "GLUT/glut.h"
 
 namespace LightmapBaker
 {
@@ -10,7 +12,7 @@ namespace Renderer
 	{
     public:
         std::vector<glm::vec3> vertices;
-        std::vector<int> indices;
+        std::vector<GLubyte> indices;
         glm::vec3 normal;
         int patchLevel;
         int elementLevel;
@@ -20,7 +22,7 @@ namespace Renderer
 
     public:
         Mesh();
-        Mesh(const std::vector<glm::vec3>& vertices, const std::vector<int>& indices, const glm::vec3& normal, const int& patchLevel, const int& elementLevel, const float& area, const glm::vec3& reflectance, const glm::vec3& emission);
+        Mesh(const std::vector<glm::vec3>& vertices, const std::vector<GLubyte>& indices, const glm::vec3& normal, const int& patchLevel, const int& elementLevel, const float& area, const glm::vec3& reflectance, const glm::vec3& emission);
         void Initialize();
         void Render();
         void Destroy();
